[logging] Don't debuglog_create(ZX_HANDLE_INVALID, ...) in core-tests
Use the root resource create a debuglog since it's typically available
when running core-tests anyway. In the failure case (where someone has
manually run core-tests, rather that running it from userboot), push a
the error message onto the stack and abort() so that it's visible in the
backtrace from crashlogger.

(I'm a bit concerned that it's pretty awkward to get the root resource
for logging even in this absolute simplest case, but we'll see how it
goes I guess.)

ZX-2184 #comment [logging] Don't debuglog_create(ZX_HANDLE_INVALID, ...) in core-tests

Test: CQ
Change-Id: Id3be962e834227f57c714ccfa4a1991ca2313ad2