[virtcon] Improper handling if there are 0's in middle of HID report.
VNC/KVM on a NUC typing "now is the time" results in something
like this:

"noow is thee timem"

Because VNC/KVM is simulating a keyboard by often reporting 2 keys
in one packet, and as the keys are released, it leaves a 0 in the
middle of a report when a key is reported as being released. Our
code stops examining the report packet when it first encounters a 0,
causing this issue.

The fix is to read the whole packet.

ZX-1482 #done

Test: NUC KVM now works. Pixelbook continues to function as expected.
Change-Id: Iafe21b1477cf7add8d921e4da135ce2071e38494