[scenic] Add SetClipPlanesCmd.
Scenic is moving away from stencil-buffer clipping; instead
clients will specify a list of clip planes which apply to a
node and all of its children.  This CL adds the API, but does
not yet implement it (i.e. content will not yet be clipped to
the specified volume).

This relands a previous CL that broke SDK clients.  This time
we append the new Cmd to the end of the command union; this is
a workaround until we transition to xunions.

SCN-670 #comment
SCN-1225 #comment

TEST=no-op; API not yet called nor implemented.

Change-Id: Ib6403533a26c2c1decd29513deb15b40fa663d8b