[kernel][topology] Convert APIC table logic into c++.
Create a c++ library (lib/apic_tables) that contains the logic that used
to be in platform/pc/acpi.h but as a c++ class. This allows the adding
of unit tests by this CL and allows future users to provide fake acpi
tables to test their use of this logic.

Issue: ZX-3068
Test: "k ut all" on qemu and eve.

Change-Id: Icf99664caefd9596fa578aaac0774372b26fdaad