[ethernet] Fix race between FIDL and unbind
When the start/stop calls release the mutex to allow client callbacks to
run, an unbind may come in and clean things up. This can cause a crash
when the start call resumes and attempts to use the edev.

In addition, close the connection if unbind has set ETHDEV_DEAD while
FIDL dispatch was waiting for lock acquisition.

Cleaned up vestigal uses of the ETHDEV0_BUSY which is not used anymore
since FIDL dispatch is done on a single thread now.

Tested: fx shell 'for i in $(seq 1 1000); do /pkgfs/packages/netstack_integration_tests/0/test/netstack_integration_test; done'
Change-Id: I92988cab3e611841cae0d83d8e1aa30ac04d64ef