[ulib][dispatcher-pool] Add default slack arguments to Timer
Add a default argument for slack type to dispatcher::Timer::Activate.

Add a default argument for slack amount to dispatcher::Timer::Arm.

In the future, job policy will set a default slack amount for most
jobs. When that happens, the slack type used to create a Zircon Timer
will matter. Different use cases call for different types of slack so
it's important that these parameters are configurable. We using C++
default arguments to facilitate the transition. A future change will
remove the default arguments (ZX-3311).

Bug: ZX-931 #comment default slack args for dispatcher::Timer
Test: none
Change-Id: I2a12e35f24563b052a13591b387d17fd932a9969