[libc] Remove use of random() in debugserver/test_apps/thread-test.c
random() is implemented() as ++i. This thread test doesn't appear to
need a lot of Serious Crypto Randomness, so switch to rand() which is
the sort of trivial random number
https://fuchsia.googlesource.com/zircon/+/master/third_party/ulib/musl/src/prng/rand.c#10
this test would appear to be happy with.

ZX-3307 #comment [libc] Remove use of random() in debugserver/test_apps/thread-test.c

Test: CQ
Change-Id: If05f34fcef8d3e50f3ac394007d0407b1b003dd2