[cs] Implement Component Stats (cs) Tool
This CL is a first step towards implementing a cs tool for Fuchsia.

The purpose of this tool is to expose information about the runtime
component hierarchy and resource usage to developers.

The canonical view of this state is in the Hub. This tool traverses the
hub pseudo-directory structure, and extracts information about realms,
and
components.

This CL does not introduce resource reporting. Subsequent CLs will add
additional information to Hub and expose that info via cs.

Test: Manually by typing cs in shell. Automated testing coming.
Change-Id: I60648101d1b930c83434c57401150c13d46dcd31