[environment] Enforce uniqueness of environment label
This enforces in appmgr that the environment label provided through
fuchsia.sys.Environment.CreateNestedEnvironment must be unique within
the scope of the containing enviroment. This will let us rely on the
label as a unique identity for the environment for the purposes of
isolating persistent mutable storage.

CF-28 #comment

Test: fx run-test appmgr_integration_tests -t
appmgr_realm_integration_tests

Change-Id: I2b4831402ec569cc72d7a5451317e76e6b2765a6