[environment] Make EnvironmentController required in CreateNestedEnvironment
The EnvironmentController request being optional is somewhat dangerous
in that the environment created will not be cleaned up automatically if
the process which created the environment dies. In other words, it
behaves as if the process called EnvironmentController.Detach(). This
makes it so that the caller of CreateNestedEnvironment must explicitly
call Detach() if that behavior is desired, rather than getting it by
default and where it may be unintended.

Note that today there are no explicit callers of
EnvironmentController.Detach() and there do not appear to be any use
cases that actually want that behavior.

Test: fx run-test appmgr_integration_tests -t
appmgr_realm_integration_tests

Change-Id: I0acde80488e8ed59df3c7ce083584dccc71fe007