[kernel][counters] Expose kernel counters to userspace as VMOs
This exposes the kernel's internal counter data structures to
(privileged) userspace as a pair of read-only VMOs.  One covers
the (sorted) array of description structures that holds the name and type
of each counter.  The other covers the arena of per-CPU int64_t[] arrays
that parallel the order of the description structure array.

The intent is that these kernel-internal data structures will remain
a private ABI known only to a privileged service that is tightly coupled
with the kernel.  As well as these data layouts, the set of counters and
their meanings (and even the variety of counter types available) is not
meant to be any kind of stable ABI between the kernel and userland.  The
counters exist purely for kernel-specific diagnostic and logging purposes.

Example output from the new kcounter tool:
    $ kcounter kernel.handles.m
    kernel.handles.made = 2003
    kernel.handles.max_live = 317
    $ kcounter -t kernel.handles.made
    2080
    $ kcounter -v kernel.handles.m
    kernel.handles.made = 572 + 444 + 492 + 418 + 0 + 0 + 0 + 0 + 0 + 0 + 0 + 0 + 0 + 0 + 0 + 0 = 1926
    kernel.handles.max_live = max(236, 120, 308, 238, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0) = 308

Bug: ZX-3010 #comment Expose kernel counters to userspace as VMOs
Test: CQ with new kcounter-test & manual use of new kcounter tool
Change-Id: I45699cd822d7b4fe46fd5a4ab38b46841d8e3cca