[debugger] Separate out formatting options.
Moves the FormatValueOptions from console/format_value.h to a new file
in the "expr" directory. This is to move to a design where the abstract
output formatting is done in the expr directory, and the conversion to
text is done in the console directory.

There should be no behavior change, this only moves and renames the
struct.

TEST=none

Change-Id: I24ad66f564bb6bae690c7be80fb2823d63a936f6