[mediaplayer] don't pad out zero video size
This CL changes the new-ish video buffer alignment code so it
doesn't try to add padding when the video size is 0x0. This was
causing unneeded, too-small VMOs from being allocated for before
the video size is known. There was no user impact, because correct
VMOs would be allocated shortly after that.

TEST: no behavior change...tested for regressions
Change-Id: I02cbd3027604dbfa9d870807b6a20a1925439dcb