[debugger] Avoid crash on initial thread exception.
The initial thread exception on a process is a weird state where no
registers can be read or set. This makes the stepping routines more
tolerant of threads with no stacks.

Adds some help text to assist people understanding the weird initial
state of the process when this happens.

Does some minor clarification in the debug_agent.

Fixes messaging around backtraces with missing stacks (the condition was
wrong for the more descriptive help text).

TEST=manual
DX-912

Change-Id: Id9a9a425e580977ae6b7dd0020d74b89d4bac52a