[scenic][escher][test] Re-enabling VK_TESTs
Since these tests are components being run in hermetic environments, they
need to have vulkan loader injected.

Previously these tests were silently not running on hardware devices due to
not being able to load Vulkan. A more reliable way to do this could be to use
test environments in the package spec, but currently the VK tests and
non-VK tests are intermixed and splitting them apart would be a bit of work.

Test: fx run-test escher_tests
      fx run-test scenic_tests -t gfx_unittests
Change-Id: I4f7c4af42c07bbb5304491cbc239b8080d294ddc