[inspect] Add RAII inspect wrappers with strict ownership.
This change implements a facade in front of the Inspect API that
enforces strong ownership semantics for objects, properties, and
metrics.

Using this is safer than using the existing ObjectDir and ExposedObject
APIs directly, and it will be the recommended way of interacting with
inspect. Additionally, the new VMO support will fit in behind the
facade.

CF-257: #done
TEST=run_test_component inspect_unittests

Change-Id: I55ff7851e48da4e9f1daaa3881e4b6d5781ede33