[software delivery] use a grand unified go binary
The total composition of all Go binaries that make up the software delivery
subsystem result in a large on-disk footprint, as well as a large VMO
footprint due to some relative inefficiencies in Go binary production.

Making all of the runtime dependency binaries a single grand unified binary
saves another 60% over both stripping and lz4'ing all of the binaries as was
already done. This moves the disk footprint for all of software delivery to
around 4mb (slight variance by architecture), after lz4 and stripping. The
grand_unified_binary is around 9.1mb uncompressed.

Note: this change requires a hard-transition with //build, as //build
directly references the pkgsvr package in order to compute command line
arguments for loader service configuration. An integration change and //build
sibling will be created as well.

Bug: US-441 #comment grand unified binary introduced, saving 60% space
Bug: PKG-418 #comment grand unified binary introduced, saving 60% space
Bug: PKG-420 #done grand unified binary reduces overhead of amber_ctl to less than 100kb
Bug: PKG-308 #done grand unified binary reduces overhead of amber_ctl to less than 100kb
Test: Verified amber, pkgsvr, and system updates still work, as well as regular SWD test suites in CQ

Change-Id: Ic8683e1142b136d4c84d5cad8479b095dcc3f7b7