[mediaplayer] fix threading issue in SimpleStreamSinkImpl
This CL fixes a threading issue in SimpleStreamSinkImpl. Callbacks
from PayloadBuffer::Create are called on an arbitrary thread. In this
case, the callback code was written as though it runs on the main
FIDL thread: unsynchronized accesses to members of 'this' happened
as well as a callback into FIDL on the wrong thread.

This problem has been addressed by posting the callback to the main
thread.

TEST: see MTWN-215 repro steps

MTWN-215 #done

Change-Id: I9a7bacb68aa48c7cac8fccee3e99fff7fdc9b46d