[netstack] remove iostate.lastError
...by monitoring connection success in loopRead.

Note that we currently don't match Linux when sockets are reused after a
failed connect() call. This adds a manual test that demonstrates the
expected behaviour (on Linux), which now produces the output:

connect failed as expected: Connection refused
connect failed: Connection refused

and exits 1. On my Linux host, this same test exits 0.

This deviation is due to upstream implementation details, but I've
attempted to write this patch in a way that will do the right thing when
upstream allows socket reuse.

Test: fx run-image-test connect_twice_test
Change-Id: I1c43711bba8b39c8e4436038d01c30873b391e6d