[blobstats] Update blob stats script to use paths within packages
This updates the blobstats script to display the paths to blobs based
on the path name used by the package containing them. The underlying
storage charges and proportional counts are still computed on a
blob-by-blob bases, based off of their hash. This uses the blobs.json
metadata file generated by the build to determine the paths within
packages.

Also restructures the internals of blobstats to reflect that blobs
don't have a meaningful path outside the context of a particular
package, and avoids recomputing hashes of blobs to speed up the
calculations and simplify the code.

Also marks unique blobs as unique.

Test: run 'fx blobstats' on builds including prebuilt_package entries
BLD-342 #done

Change-Id: I0a6a8d40ae8ad77d362068b5fc58fcd7390b83d2