[vmo] Avoid assert for very large vmo allocations
When a VMO was allocated in the range at the end of the address space, a
later attempt to free by range would overflow the end bound, resulting
in the last page(s) not being freed.

Reduce the maximum size of a VMO so that a full VmPageListNode's span
always has an offset that fits in a uint64_t.

This bug was found by syzkaller. Minimized repro from cpu's CL at
https://fuchsia-review.googlesource.com/c/zircon/+/184607/2:

void crasher(void) {
    const uint64_t kVmoSz = 0xffffffffffff7fff;
    zx_handle_t vmo;
    zx_status_t res = zx_vmo_create(kVmoSz, 0, &vmo);
    if (res != ZX_OK)
        return;
    zx_vmo_op_range(vmo, ZX_VMO_OP_COMMIT, kVmoSz - 0x8000, 3, 0, 0);
    zx_handle_close(vmo);
}

ZX-2359 #comment [vmo] Avoid assert for very large vmo allocations

Test: new cases added to `k ut vm` and `k ut vmpl`
Change-Id: I799ef6cfd1629f4a259fd1746d0b0d4e1ef3e32d