[kernel][mutex] Make mutex_t use constructors/destructors.
Take a baby step in the direction of making mutex_t more C++ and less
C.  Specifically...

++ Remove mutex_init
++ Remove mutex_destroy
++ Remove the MUTEX_INITIAL_VALUE { .member = ... } initializer macro
++ Wire up a constexpr default constructor which does what mutex_init
   and MUTEX_INITAL_VALUE used to do.
++ Wire up a destructor which does what mutex_destroy used to do.

I have verified (by manual inspection) that our current debug builds
(for clang and ASAN) are still performing initialization of globals
and static locals using just .data segment initialization.  There is a
separate discussion going on about how to ensure that this behavior
remains safe moving forward from the static local perspective.

This change is a small step in the direction of allowing us to turn
this code into a full C++ implementation (from a style perspective).
Eventually, we can use this as part of the process to to remove any of
the #ifdef KERNEL stuff from fbl::AutoLock, and to migrate the
kernel's RAII lock object over completely to lockdep::Guard instead.

Tests: Built for x64 and arm64 (ASAN builds).
       "k ut all" and "k thread_tests" in QEMU for both architectures.
Change-Id: Ic97d83c790d4297fabb0997baf0b1d9bc94fa9b6