[fonts] Update to use CIPD prebuilt
This is the start of a soft transition to use the CIPD prebuilt. Once
the change to the ensure file and the noop update script make it to GI,
another change to integration manifest will remove the Jiri hook. After
that, the upload.sh and update.sh scripts will be removed.

Test: Builds locally, CQ
Change-Id: Ib2d761cb4578d0baf673364d8a46df77d2e008d2