[modular][storymodel] Un-flake ledger_story_model_storage_unittest.
The old test was flaky because it depended on the Ledger notifying the
StoryModelStorage of changes to establish what would be the expected
value of a later read from the Ledger, but it did not wait on a
condition that guaranteed the Ledger had notified it. This created a
race (against TIME!).

TEST=ledger_story_model_storage_unittest

MF-165 #done [modular][storymodel] Un-flake ledger_story_model_storage_unittest.
MF-165 #comment

Change-Id: I0d515609374922692caf2fc46491e52d7354a158