[sl4f] send back errors in sl4f
When executing commands in sl4f, we need to return the errors
encountered to the device driving the test.  This CL captures the error
and sends it back over to the test driver as an AsyncResponse.

This CL also fixes an existing issue that would trigger a panic if the
dut is unable to send back the response.  If there is an error returning
the response, it will be silently dropped as there is little the dut can
do (the connection to the server may be broken).

Bug: CONN-3
Test: act.py -tc WlanScanTest (with and without a wlan interface to
trigger an error)

Change-Id: I6a87eef54a2224f23db1f8f386308b643610b890