[debugger] Expand inline frames.
When an instruction pointer is inside an inline frame, this patch will
expand the inline chain to multiple stack frames. The result is that the
list in the "backtrace" and "frame" commands much better represents the
programmer's understanding of the code.

Thread control commands have not yet been updated to understand these
inline frames. As a result, step, next, and finish will do unexpected
things when these inline frames are encountered. However, they didn't
work the way users expected before this patch either so this is not a
notable regression. These commands will be updated in later passes.

Add inline function call location to the Function symbol.

TEST=unit,manual

Change-Id: Ib709c62a482605387dfc0b2e3591b282750b2126