[dev][operation] Add operation library
This library provides a generic solution to the private section problem
which exists for types such as usb_request_t, node_operation_t,
block_opt_t, and others. Specialized wrappers for each of those types
can be built on top this library.

More specifically, the problem exists whereby a series of drivers reuse
the same object as it traverses the driver stack for a specific
subsystem. There exists a public section specified by a banjo protocol,
along with a private section for each layer in driver stack appended to
the end of it like so:

---------------------
| Public Definition |
---------------------
| Driver 1 Private  |
---------------------
| Driver 2 Private  |
---------------------
|        ...        |
---------------------
| Driver N Private  |
---------------------

Driver N in this case would perform the allocation of the entire struct.
Driver 1 in the example above would be the device driver which talks
directly to hardware. The request would only be "owned" by a single
driver at a time, but only Driver N (the one who allocated the request)
would be allowed to free it.

The library is inspired by the usb::Request and friends library. They
will transition to being built on top of this library in future CLs.

Tested: `runtests -t operation-test`
Change-Id: I3ab6bdab14ef5f1654bde9817321286bf5e6eb36