[biscotti] Remove legacy service endpoints.
We initially published view providers both to the root service
directory and to the public service directory. This was because appmgr
was in a transition from the flat structure to the public directory
structure.

Now that appmgr is looking into the public directory, the legacy root
serice entries can be removed.

Test: tiles_ctl add linux://chromium
Change-Id: I7a697495e7799fc1018d0cf3f05667a95a965e60