[guest][integration] Wait for guest to finish initial serial writes
Attempting to use the serial socket whilst the guest is still generating output will
cause commands to spuriously fail and have to retry. This improves the performance of
the tests in some situations by waiting until no data is seen on the socket for a
timeout in order to guess that the guest is 'finished'.

Test: run-test guest_integration_tests

Change-Id: I00a9501550c1f971ef891745ce771d9d5e418829