[zircon][hid-parser] Add ability to extract values with units
Our ExtractUint functions are returning the raw data from the
report. This is not enough for a device that expects to be
able to read the report information in the context of a certain
unit.

ExtractWithUnit takes care of doing sign extensions and
logical to physical value conversions so the value it returns
actually represents the Unit outlined by the HID report descriptor.

This also fixes a bug where Max physical/logical values need
to be unsigned if the Min value is greater than or equal to zero.

TEST: fx shell runtests -a -t hidparse-test
Change-Id: Id19b466cffd97d7ba5b132ae48d87c07ccfdd2f4