[media][codec] Copy Codec to StreamProcessor
The FIDL definition of the Codec interface can be used to represent a
more generic interface. At its core, the Codec interface is essentially
an asynchronous data processor. In various media pipelines, it is likely
that we will have multiple data processors and it would be ideal to
re-use the same interface for all. This will allow players to handle
them generically and easily pipe outputs to the next stage's inputs.

This CL does nothing more than copy codec.fidl and codec_common.fidl to
new files in the fuchsia.media library and rename the files to
stream_processor.fidl and stream_common.fidl, respectively. Subsequent
CL will actually rename the structure within them.

Test: None, as the fidl sources are not yet added to build files.
Change-Id: Ide9105f963ddb5aa4c36475c7c7fc7e79b26dfcc