[media][codec] Rename Codec to StreamProcessor
The FIDL definition of the Codec interface can be used to represent a
more generic interface. At its core, the Codec interface is essentially
an asynchronous data processor. In various media pipelines, it is likely
that we will have multiple data processors and it would be ideal to
re-use the same interface for all. This will allow players to handle
them generically and easily pipe outputs to the next stage's inputs.

This CL renames Codec to StreamProcessor and begins the transition.

Test: FIDL compiles. No behavior changes.
Change-Id: I296411d27d2ca181ff1fd36f3948c22607ae2b8b