[component_manager] Component manager scaffold.
This change introduces the beginning of the component manager.  It
supports a very simple component topology consisting only of a root
component instance.  The component resolver and runner interfaces
are stubbed out; they will be implemented in future patches.
Plenty of stuff left to do!

And thus the hippos waded into the bog, eager to greet their
new friends...

Test: run fuchsia-pkg://fuchsia.com/component_manager_tests#meta/component_manager_tests.cmx
Test: run fuchsia-pkg://fuchsia.com/component_manager#meta/component_manager.cmx fuchsia-boot:///boot#meta/scaffold.cm
Bug: CF-154
Change-Id: Ia70d76cae776a03ff9c418f210e34686c88ca808