[ledger][benchmarks] Randomize keys.
Keys generated for the benchmarks had the index as a prefix, making
them partially sequential (e.g. 1, 10, 2, 3, 4, 5, 6, 7, 8, 9). After
this CL, keys will have the index as a suffix, forcing a random order
in insertions/updates, while still allowing retrieving the keys' ids.

Note that this change will likely change some of the results in the
dashboard, because of the re-ordering of the keys.

Test: All benchmarks are still running correctly.
Change-Id: If88eeb556dbccf1a668d3af2040614d9b068d7b1