[drivers][usb-mass-storage] Fix race condition in USB mass storage driver
Currently; it is possible for a race condition to occur in UMS between the call to sync_completion_wait
and sync_completion_reset which could result in the last transaction getting dropped if no others
come after it.

This patch checks for an empty list instead of a timeout from sync_completion_wait to
avoid such behavior.

TEST: Validate that UMS still works (unit tests coming in my refactoring CL)
Change-Id: I924e5c2b10ac93649379accba2751360d535bcb3