[media][codec] Convert Codec to StreamProcessor
The FIDL definition of the Codec interface can be used to represent a
more generic interface. At its core, the Codec interface is essentially
an asynchronous data processor. In various media pipelines, it is likely
that we will have multiple data processors and it would be ideal to
re-use the same interface for all. This will allow players to handle
them generically and easily pipe outputs to the next stage's inputs.

This CL converts most usages of mediacodec.Codec to the new
media.StreamProcessor equivalent in garnet. As this is a soft
transition, the Codec interface has been left in place and the
CodecFactory will still support the old interface. The CodecImpl library
has been updated to support the new StreamProcessor interface and there
is a temporary binding wrapper that will support using either a Codec
binding or StreamProcessor binding. This should be removed after the
transition is complete.

This CL also includes a change to the CreateDecoder_Params of the
initial format details to be of fuchsia.media.FormatDetails rather than
CodecFormatDetails. The two structs remain binary compatible and the
change of type should not break the one out of tree usage of
CreateDecoder_Params.

Test: mediaplayer_test_util and use_media_decoder on vim2

Change-Id: I42b6736ecfc493d4249fa52b29d6aa9b8cc965df