[ulib][zxcpp] Implement aligned delete operators
C++17 added a number of delete operators, including versions
which take the size of the object to delete and/or the alignment
of the object to delete.

Compilers in c++17 mode (both gcc and clang for zircon) may emit
calls to these operators - empirically gcc on aarch64 in no-opt
mode does so with a particular commit (231861) included.

Implement the align/size+aligned delete operators in terms of free().
This is safe - aligned allocations via new ultimately just call
aligned_alloc().

Tested:
Allows building https://fuchsia-review.googlesource.com/c/zircon/+/231861
Note this is only to allow building; it is not to intended to make the
tests in that commit pass.

See: https://fuchsia-review.googlesource.com/c/zircon/+/243712 for the
commits folded together.

Change-Id: I9b12609af5385602d7ac69221e3a8d4cc82d8f73