[sessionmgr] Cleanup while reading the code.
Rearrange the construction of the service namespace passed to the
session shell such that the registration of the connection factory
and the declaration of the service name are adjacent for each
service in the service namespace.

Use type inference for the factory function |request| arguments,
as they are determined by the type parameter of the AddService<>
invocation.

TESTED=/pkgfs/packages/modular_tests/0/test/run_modular_tests.sh

Change-Id: I357284e6df33b0e81aa956b43bad2b03a7ed80a6