[Terminal] Now takes keyboard input.
The terminal has now been plugged in to IME to accept keyboard input
events. For the moment, it forwards them directly into the terminal
model, because it doesn't have shell integration yet.

Also update the `HostImageCycler` to stop attempting to re-use the
buffer each frame. It seems that Scenic doesn't properly support this
yet, which meant the terminal couldn't update it's display each frame.
JIRA MZ-145 will address this. Thanks to Rob Tsuk for figuring this out.

Test: Manually verified, in Ermine on a PixelBook.
Change-Id: I540dfdf0ec1fab4212f66a5375fe1054e8a2401e