[wlan] Decommission CBW_COUNT enum variant across layers
This is a clean-up CL. CBW_CONUT was originally introduced
to assist sanitization of the CBW value. Its use added no value
but more inconvenience, leading to unclean Rust side API signature.

CBW is not a value coming from the air. The use of CBW_COUNT
in WLAN device drivers do not exist.

Test: Build.
Change-Id: I6013b4f8529df0446e36e3e97cd593386db2fe1c