[upci] Implement bus enumeration
This change ports the kernel PcieDevice and PcieBridge classes over to
userspace as Device and Bridge, respectively. Many of the original
concepts were maintained, but bus/device ownership and scanning
methodology has changed to suit a world where the Bus driver owns a
given set of bus ids and must handle bridges from there. The Bus driver
finds valid config space entries and creates bridges and devices
corresponding to them, keeping track of bridges so it can iteratively
BFS the PCI bus topology and build a graph. This will need to change to
DFS to be proper enumeration to spec in a subsequent CL, but it is
sufficient to test the graph lifecycle. These scans are bounded by the
ranges reported to the Bus driver through the PciRoot protocol.

On boot the bus driver will scan all devices, allocate them, link
them to the graph, and then request the root disable/unplug all
downstream devices as a test.

Bridge window and device bar allocations are not handled in this CL and
will come later.

ZX-3146

Test: Verified that all devices/bridges are found, linked, and released
without any leakages or asserts with the dtors and data structure
checks.

Change-Id: Ic121b5363c92ec35f0b61cf78427ab162d28a1f7