[blobfs] Teach blobfs host tool to record the size of blobs as it writes
This adds a '--sizes' option to  the blobfs host tool which tells the
tool to write down the amount of space used by each blob after any
processing it may perform internally (compression, alignment, etc) has
been done. The output format is manifest style, that is a list of lines
each of the form:

<blob merkle digest>=<number of bytes>

This can be used by tools higher up the stack to get a more accurate
picture of what's going on inside a blobfs partition.

Test: run blobfs tool with '--sizes <filename>', examine output
ZX-3303 #comment host tool support

Change-Id: Ifcdd7c7305123713e0dfdb456292635cbe73fb16