[ledger] Fix use-after-free in PageManagerContainer
PageManagerContainer keeps a token on itself during internal requests to
ensure it is not destroyed before reaching the end of SetPageManager.
The method PageIsClosedAndSatisfiesPredicate relied on the ability to
close the page by calling the token. We cannot rely on that anymore: we
post a task so that the call to SetPageManager finishes properly before
we continue.
We need to ensure that the LedgerManager stays alive between the end of
the SetPageManager call and the call to the coroutine: CheckEmpty now
checks that we are not tracking a page.
The LedgerManager may also be deleted because Ledger shut down. We use a
weak pointer to check for that.
To make things safer, the page tracking code now generates a token
that is removed on destruction, and we use callback::EnsureCalled to
ensure that we always call the callback.
Also fixes a similar problem where PageConnectionNotifier could
recursively call CheckEmpty() and get destructed before the end of the
outermost CheckEmpty() call. This is done by moving the ExpiringToken
into the PageConnectionNotifier (the logic only touched the notifier)
and taking a token before calling OnPageUnused.

TEST=passes asan for all unit and integration tests

LE-668 #done
LE-684 #done

Change-Id: I7945b6eeac8ae4012eb75fe2c7ccfa1a1dbf012d