[debugger] Initial watchpoint debug agent support.
This is the infrastructure plumbing to get watchpoint going.
This is mirroring very closely the breakpoint implementation, but
purposedly being separated. There are key details of the watchpoint that
makes it not use ProcessBreakpoint amd most probably will divert further
over time.

Created unittests that verify that the installation is calling the
underlying arch calls.

TEST=unit

Change-Id: I95116a7d0d43196377365765550a8b1476ca02bb