[virtio_wl] Only one command per descriptor.
The linux driver will only parse a single cmd header from a descriptor. This
causes issues when sending the XKB keymap to a client since the virtio_wl
device will place the NEW_VFD command for the keymap VMO in the same
descriptor as the RECV command for the wayland message. The result is that
the guest driver will create the new VFD, but the event message that
references the new VFD will be dropped.

Test: Verified the bridge can send a keymap into the guest. This enables
      keyboard input on biscotti.
Change-Id: I0030b0cdcfc86e8dda6b28df02cf27f86fb5887a