[fidlc] Refactor tables generator, and add unit tests
We split the TablesGenerator into a CodedTypesGenerator which does smart
things (calculates coded types), and the TablesGenerator which emits
code (that part is expected to be straightforward string manipulation).

This split makes it possible to test the CodedTypesGenerator more
easily, and clarifies the surface which needs to be tested.

To reduce the diff size, making the TablesGenerator extend the CodedTypesGenerator
(vs delegating to it). We may want to refactor that separately, but
think this is fine as is.

Test: make tools -j12 HOST_USE_ASAN=true && ./build-x64/host_tests/fidl-compiler-test
Change-Id: I407892448ba706012456a99b2ab03c9e991fb584