[blobstats] Fork blobstats visualization template, improve
This forks the visualization template and dependencies of the blobstats
tool and updates them to take into account features of blobs and
packages, instead of pretending everything is a symbol in a binary.

Improvements:
* Update all terminology (symbol->blob, directory->package)
* Add information about compressed and uncompressed sizes and reference
count on shared blobs
* Update colors to be more neutral
* Update labels on entries to be more accurate
* Update report naming and contents
* Remove dead links
* Add original and estimated compressed sizes to data model

Also moves the code to its own directory and imports the d3 library
instead of reaching into the Dart VM's internal directory.

Test: fx blobstats

Change-Id: I95cea2fa7e4c005d0917c991a726098326b00116