[cs] Add ability to query inspect objects of components
This CL introduces the command line argument(s): --inspect or -i
followed by a job ID.

It will traversal the component hierarchy, finding all components in
that job and print out information about all the threads in the job via the
Inspect API.

Note that this can be a massive spew. A subsequent CL will clean up the
UI further.

Change-Id: Ia7a1b8dce3f04017c516459ee10041e557f38dc6
Test: manually cs -i <job_id>