[pkg_resolver] temporarily disable tuf
We are doing a ring upgrade to 0.14, which is a breaking change.
Since we aren't using tuf yet, to skip going through the whole
breaking change process, this CL temporarily removes the tuf
dependency from pkg_resolver. Once the ring and tuf update change
lands, we will revert this patch.

Test: CQ
Change-Id: Ia7e0407965298f18369f6f8badcdcb3ce28adbd4