[mt-usb] correctly set up fifo buffers
Interleaved IN/OUT transactions were sometimes failing due to a misconfiguration
of the FIFO buffers. When double-packet buffering is enabled, the programmed
FIFO size should be encoded as the size of only one of those buffers, not both.
This change modifies the FIFO configuration; the previous setup was causing
buffer overlaps and spurious garbage transmissions as a result. This new setup
keeps all of the buffers properly aligned and isolated.

Test: paving

Change-Id: If0d4a6bb53559d3eb8b3e7200502be7802b79604