[cobalt] SystemMetricsDaemon: New version.
We have written a new version of the SystemMetricsDaemon with the
following changes:

(1) We use new Cobalt 1.0 locally aggregated metrics.
(2) We reorganized the code into multiple files.
(3) We reorganized the code to make it more testable.
(4) We added extensive testing.

In support of this we also introduce a new cobalt/testing library
containing the following:

(1) FakeSteadyClock
(2) FakeLogger_Sync

TEST=unit test and manual test

Change-Id: I3cb7038ab4ed0a02a32d18e6a24b360b86f24944