[debugger] Add inline frames to fingerprints.
FrameFingerprints are used to compare stack frames. The identifier used
is the previuos frame's stack pointer, but this fails for inline frames.

This patch adds an inline function count to the frame fingerprint so
that all frames can be compared for ordering whether or not they're
inline frames.

These comparisons are used as part of the thread controllers for
stepping. It may change stepping in the presence of inline frames, but
the stepping still needs to be updated to handle inline frames
correctly so this patch won't make it work.

TEST=unit

Change-Id: Iaa8c439c6d3c02e285ebb58c6ed7ceb0edc52e0a