[debugger] Enhance address range handling.
Creates a dedicated class for handling a set of address ranges. This
gives one place to implement querying inclusion for discontiguous
address ranges.

Use this to represent code ranges for function objects instead of the
existing vector. This should have equivalent functionality but provides
more tolerance of weird inputs from the DWARF files.

Use the new ranges to represent step ranges. This is not currently used
but in the future will allow stepping in more complex things like inline
functions that can be represented by multiple discontiguous ranges.

TEST=unit

Change-Id: Ifb99dc768e79bc66f61ce00150a3c71e873f95d6